Added Initial Parts for Pickers
Added the layout for Pickers (Date and Time).

Implemented a mobile and desktop version of the date picker using the moment library in a dialog.

The date picker will appear in landscape if it can fit, otherwise it will appear vertically.

The current functionality is only to switch the month and select a date. You are unable to set that value in the text field.