Enhancement/#25 select field selection (#69)
* Menu Performance Optimizations

Updated the Menu component to only have a single click event listener. This allows for speedier
updates since we won't need to bind x amount of onClick listeners. There was also more of a chance
for the pure render mixin to consider the onClick functions not equal.

* Updated ListTile for Tabbability

Updated the ListTile to have a tabIndex of 0 by default.

* Implemented #25

The select field now allows for a new item to be selected by pressing a letter or number
to select an item beginning with that letter or number.

The user can press the up or down arrow key to select the next or previous item.

Pressing tab will advance to the next tabbable item on the page and not the menu items.

* Added Select Field Tests #52

Updated to have SelectField test and fixed some bugs discovered while testing.

Fixed the cursor for disabled select fields.

Fixed cursor for buttons on hover.