Feature/navigation drawer rehaul (#96)
* Added a new DrawerHeader Component

This component is almsot the same as the old NavigationDrawerHeader component.

The only differences are the new one uses the PureRenderMixin and some small fixes to work
with different media sizes.

* Created a Drawer Component

Extracted out the sliding Drawer nav from the NavigationDrawer into it's own component.
This was done because the drawer will only be rendered when open now and switching
to using the CSSTransitionGroup for appearing/disappearing.

* Added the DrawerToolbar Component

This is basically the same as the NavigationDrawerToolbar component. just a few changes
for the responsive media changes in the new NavigationDrawer component that is coming up.

* Added the new NavigationDrawer component

Added the new NavigationDrawer component that will listen to media size changes and update
the display style appropriately. It is currently named _NavigationDrawer because I need
to add deprecation warnings for this current release and then delete the old in the next release.

* All the CSS Changes for NavigationDrawer Rehaul

This is all the css changes to get the navigationd rawer to be more responsive at
differnet media sizes. It breaks the old drawer.

* Revert "All the CSS Changes for NavigationDrawer Rehaul"

This reverts commit a11e775b8f4676c5d8566220412328f59bedbd2c.

* Added Deprecation Warnings and UpgradeGuide

Added deprecation warnings and an upgrade guide for the navigation drawer component.

* Fixed DrawerToolbar test I think