Updated DatePicker CalendarDates
The Date Picker now allows for an initialCalendarDate prop which
will set the date the calendar should initially open to the first
time the component is opened or any consectutive times if the user
never hits OK.

Also updated the calendarDate and calendarTempDate to be modified
by default based on the min/max dates. This will not allow the
calendar to be opened outside of the min/max date range. It also
updates when the min/max date prop changes.