Added Portal Component and Small Button CSS
Added a Portal component so that some components do not actually get
rendered until they are completely visible. It will create another react
subtree outside of the main tree by creating it as the last body
element.

Converted the Dialog Component to use the Portal.

Small padding fix for text buttons so the text is centered.