Fixed the Documentation Website for #177
The documentation website was _flashing_ after initial page load because
the styles that were not in the main chunk were being injected into the
head after initial page load. The Fonts and material-icons were also
only being included after initial page load.

I fixed the build steps to make sure that all styles are included in the
main css bundle and moved the font libraries to the <head>.