Added Autocomplete Search Example
Added an example of using the Autocomplete to make an application where
the main focus is searching. It also showcases the uses of the
simplifiedMenu prop.

This also updated the custom-sizes mixin for text fields so that only
one item is required in the custom-size map when not using floating
labels.

Also fixed the toolbartitle to not erase existing ids for the title
prop.

Finally, added a component to conditionally render components based on
media.