Export the Component from typescript definition files
When using refs (e.g. on TextField), the Component type is the
type of the reference. In order to be able to run methods on
the referenced component, the type must be exported.

```
import TextField, { TextFieldComponent } from 'react-md/lib/TextFields';

// ...

  <TextField ref={(ref: TextFieldComponent) => ref.focus()} />
```

If the component type is not exported, it's not possible to call
methods in a typesafe manner.